filter visual studio dlls
Version 3.12.4 worked but included references to `
System.Text.Encodings.Web.dll` and
`Microsoft.VisualStudio.Extensibility.Editor.Contracts.dll` when
 the VSIX was rebuilt in the latest VS 2026.  This could create problem
if installed on an instance of Visual Studio that contained references
to those same libraries but different versions of the files.  This
release updates the project to prevent those files from being copied.